+++ src/app/core/graph-options/graph-options.component.ts
@@ -240,6 +240,7 @@ export class GraphOptionsComponent implements OnInit {
     this.graphSelectionsService.flagsSubject.subscribe((flags) => {
       if (flags) {
         if (flags.length > 0) {
+          this.flaggedData = flags;
           this.disableEnable('flagBtn', true, true);
         } else {
           this.disableEnable('flagBtn', false, true);